Changed Integer field to Float Field

+++ forms.py
@@ -1,13 +1,13 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, SubmitField, IntegerField, SelectField, PasswordField
+from wtforms import StringField, SubmitField, IntegerField, SelectField, PasswordField, FloatField
 from wtforms.validators import DataRequired, Email, EqualTo
 
 
 class AddProduct(FlaskForm):
     product_name = StringField("Product Name", validators=[DataRequired()])
     product_quantity = IntegerField("Product Quantity", validators=[DataRequired()])
-    purchase_price = IntegerField("Purchase Price", validators=[DataRequired()])
-    selling_price = IntegerField("Product Price", validators=[DataRequired()])
+    purchase_price = FloatField("Purchase Price", validators=[DataRequired()])
+    selling_price = FloatField("Product Price", validators=[DataRequired()])
     submit = SubmitField("Add Product")
 
 
@@ -18,8 +18,8 @@ class Update(FlaskForm):
 
 class Restock(FlaskForm):
     quantity_brought = IntegerField("Quantity")
-    purchase_price = IntegerField("Purchase Price")
-    selling_price = IntegerField("Selling Price")
+    purchase_price = FloatField("Purchase Price")
+    selling_price = FloatField("Selling Price")
     submit = SubmitField('Restock')
