+++ main.py
@@ -44,22 +44,22 @@ class Product(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(100), nullable=False, unique=True)
     total_quantity = db.Column(db.Integer, nullable=False)
-    purchase_price = db.Column(db.Integer, nullable=False)
-    selling_price = db.Column(db.Integer, nullable=False)
+    purchase_price = db.Column(db.Float, nullable=False)
+    selling_price = db.Column(db.Float, nullable=False)
     date = db.Column(db.String(100), nullable=False)
     quantity_sold = db.Column(db.Integer, nullable=False)
     quantity_left = db.Column(db.Integer, nullable=False)
-    amount_sold = db.Column(db.Integer, nullable=False)
-    total_amount = db.Column(db.Integer, nullable=False)
+    amount_sold = db.Column(db.Float, nullable=False)
+    total_amount = db.Column(db.Float, nullable=False)
 
 
 class SoldItem(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     date = db.Column(db.String(20), nullable=False)
     name = db.Column(db.String(80), nullable=False)
-    price = db.Column(db.Integer, nullable=False)
+    price = db.Column(db.Float, nullable=False)
     quantity_sold = db.Column(db.Integer, nullable=False)
-    amount = db.Column(db.Integer, nullable=False)
+    amount = db.Column(db.Float, nullable=False)
     status = db.Column(db.String(40), nullable=False)
 
 
@@ -67,9 +67,9 @@ class Received(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     date = db.Column(db.String(20), nullable=False)
     name = db.Column(db.String(80), nullable=False)
-    price = db.Column(db.Integer, nullable=False)
-    quantity_sold = db.Column(db.Integer, nullable=False)
-    amount = db.Column(db.Integer, nullable=False)
+    price = db.Column(db.Float, nullable=False)
+    quantity_sold = db.Column(db.Float, nullable=False)
+    amount = db.Column(db.Float, nullable=False)
     status = db.Column(db.String(40), nullable=False)
